fix errors and warning

+++ src/main/java/com/jhlabs/image/CircleDownGradientBlend.java
@@ -1,10 +1,8 @@
 package com.jhlabs.image;
 
-import java.awt.Color;
-
 public class CircleDownGradientBlend implements GradientBlend{
 	@Override
-	public int GradientBlendOperation(int t){
+	public float GradientBlendOperation(float t){
 		return 1 - (float) Math.sqrt(1 - t * t);
 	}
 }
