+++ src/main/java/com/jhlabs/image/GradientBlend.java
@@ -1,8 +1,6 @@
 package com.jhlabs.image;
 
-import java.awt.Color;
-
 public interface GradientBlend {
 
-    public int GradientBlendOperation(int t);
+    public float GradientBlendOperation(float t);
 }
