+++ src/main/java/com/jhlabs/image/LinearGradientBlend.java
@@ -1,10 +1,8 @@
 package com.jhlabs.image;
 
-import java.awt.Color;
-
 public class LinearGradientBlend implements GradientBlend{
 	@Override
-	public int GradientBlendOperation(int t){
+	public float GradientBlendOperation(float t){
 		return t;
 	}
 }
